requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
APScheduler==3.10.1
qrcode==7.4.2
Pillow==9.4.0
pandas==2.2.3
openpyxl==3.1.2
reportlab==4.0.4
Werkzeug==2.3.7
gunicorn==21.2.0
